requirements.txt
colorful==0.5.4
dictquery==0.4.0
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
flake8==3.9.0
future==0.18.2
logzero==1.6.3
mccabe==0.6.1
numpy==1.20.2
opencv-python==4.5.1.48
prettyprinter==0.18.0
pycodestyle==2.7.0
pyflakes==2.3.0
Pygments==2.8.1
